Remove flux emoji store
