+++ actions/emoji_actions.jsx
@@ -3,13 +3,12 @@
 
 import * as EmojiActions from 'mattermost-redux/actions/emojis';
 
-import {getEmojiMap} from 'selectors/emojis';
-
 import AppDispatcher from 'dispatcher/app_dispatcher.jsx';
 import store from 'stores/redux_store.jsx';
-import EmojiStore from 'stores/emoji_store.jsx';
+import {setGlobalItem} from 'actions/storage';
+import {getEmojiMap, getRecentEmojis} from 'selectors/emojis';
 
-import {ActionTypes} from 'utils/constants.jsx';
+import {ActionTypes, Constants} from 'utils/constants.jsx';
 
 export async function addEmoji(emoji, image, success, error) {
     const {data, error: err} = await EmojiActions.createCustomEmoji(emoji, image)(store.dispatch, store.getState);
@@ -43,7 +42,7 @@ export function loadRecentlyUsedCustomEmojis() {
             return {data: true};
         }
 
-        const recentEmojis = EmojiStore.getRecentEmojis();
+        const recentEmojis = getRecentEmojis(getState());
         const emojiMap = getEmojiMap(getState());
         const missingEmojis = recentEmojis.filter((name) => !emojiMap.has(name));
 
@@ -62,5 +61,37 @@ export function incrementEmojiPickerPage() {
         });
 
         return {data: true};
+    };
+}
+
+const MAXIMUM_RECENT_EMOJI = 27;
+
+export function addRecentEmoji(alias) {
+    return (dispatch, getState) => {
+        const recentEmojis = getRecentEmojis(getState());
+        const emojiMap = getEmojiMap(getState());
+
+        let name;
+        const emoji = emojiMap.get(alias);
+        if (!emoji) {
+            return;
+        } else if (emoji.name) {
+            name = emoji.name;
+        } else {
+            name = emoji.aliases[0];
+        }
+
+        const index = recentEmojis.indexOf(name);
+        if (index !== -1) {
+            recentEmojis.splice(index, 1);
+        }
+
+        recentEmojis.push(name);
+
+        if (recentEmojis.length > MAXIMUM_RECENT_EMOJI) {
+            recentEmojis.splice(0, recentEmojis.length - MAXIMUM_RECENT_EMOJI);
+        }
+
+        dispatch(setGlobalItem(Constants.RECENT_EMOJI_KEY, JSON.stringify(recentEmojis)));
     };
 }
