+++ components/post_emoji/index.jsx
@@ -5,8 +5,8 @@ import {connect} from 'react-redux';
 
 import {getConfig} from 'mattermost-redux/selectors/entities/general';
 import {getCurrentUserId} from 'mattermost-redux/selectors/entities/users';
+import {getEmojiImageUrl} from 'mattermost-redux/utils/emoji_utils';
 
-import EmojiStore from 'stores/emoji_store.jsx';
 import {getEmojiMap} from 'selectors/emojis';
 
 import PostEmoji from './post_emoji.jsx';
@@ -18,7 +18,7 @@ function mapStateToProps(state, ownProps) {
     let imageUrl = '';
     let displayTextOnly = false;
     if (emoji) {
-        imageUrl = EmojiStore.getEmojiImageUrl(emoji);
+        imageUrl = getEmojiImageUrl(emoji);
     } else {
         displayTextOnly = state.entities.emojis.nonExistentEmoji.has(ownProps.name) ||
             getConfig(state).EnableCustomEmoji !== 'true' ||
