+++ selectors/emojis.js
@@ -4,11 +4,50 @@
 import {createSelector} from 'reselect';
 import {getCustomEmojisByName} from 'mattermost-redux/selectors/entities/emojis';
 
-import {EmojiMap} from 'stores/emoji_store.jsx';
+import Constants from 'utils/constants.jsx';
+import {getItemFromStorage} from 'selectors/storage';
+import EmojiMap from 'utils/emoji_map';
 
 export const getEmojiMap = createSelector(
     getCustomEmojisByName,
     (customEmojisByName) => {
         return new EmojiMap(customEmojisByName);
     }
+);
+
+export const hasSystemEmoji = createSelector(
+    getEmojiMap,
+    (emojiMap) => {
+        return emojiMap.EmojiIndicesByAlias.has(name);
+    },
+);
+
+export const getRecentEmojis = createSelector(
+    (state) => state.storage,
+    (storage) => {
+        if (!storage || !storage.storage) {
+            return [];
+        }
+
+        let recentEmojis;
+        try {
+            recentEmojis = JSON.parse(getItemFromStorage(storage.storage, Constants.RECENT_EMOJI_KEY, null));
+        } catch (e) {
+            // Errors are handled below
+        }
+
+        if (!recentEmojis) {
+            return [];
+        }
+
+        if (recentEmojis.length > 0 && typeof recentEmojis[0] === 'object') {
+            // Prior to PLT-7267, recent emojis were stored with the entire object for the emoji, but this
+            // has been changed to store only the names of the emojis, so we need to change that
+            recentEmojis = recentEmojis.map((emoji) => {
+                return emoji.name || emoji.aliases[0];
+            });
+        }
+
+        return recentEmojis;
+    },
 );
